Compute optimal binary search trees
Construct an optimal binary search tree for elements with given access
probabilities.  The nodes will be labeled 1 through n for simplicity.
For the same reason, we assume that we will only search for things
contained in the tree.  The implementation uses a naive dynamic
programming algorithm that runs in O(n^3).  For a more efficient
solution, see Knuth's O(n^2) algorithm.